+++ packages/shader-lab/src/index.ts
@@ -1,5 +1,5 @@
 import { AstNodeUtils } from "./AstNodeUtils";
 export { ShaderLab } from "./ShaderLab";
 export { ShaderVisitor, parser } from "./ShaderVisitor";
-export { ShaderParser } from "./parser";
+export { ShaderParser } from "./parser/ShaderParser";
 export const parseShader = AstNodeUtils.parseShader;
